bbstat: move profile + apply into BinByBinStat
Step 4 of the bbstat refactor — the largest. Lifts the entire β profile
and application logic out of Fitter._compute_yields_with_beta (~600
lines covering 8 lite + 8 full × 3 likelihood-form branches, including
the gamma+full Newton iteration with the implicit-function-theorem
trick) into BinByBinStat.profile_and_apply.

Fitter._compute_yields_with_beta is now a 10-line delegator that calls
self._compute_yields_noBBB to get the model prediction then hands off
to self.bbstat.profile_and_apply. The Newton-loop helper
(_newton_solve_nbeta) lives on BinByBinStat and is shared between the
chisqFit and Poisson gamma+full paths.

The 6 binByBinStatType × binByBinStatMode regression on
tests/make_tensor.py reproduces the previous σ values to all reported
digits.
